+++ .eslintrc.js
@@ -42,6 +42,7 @@ module.exports = {
             "anonymous": "never",
             "named": "never",
             "asyncArrow": "always"
-        }]
+        }],
+        "arrow-spacing": ["error", { "before": true, "after": true }]
     },
 };